+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Fix `%edgeql` for modules.
+
 ## 0.1.0
 
 Initial release!